fix CMakeLists

+++ src/aarch64/mmu.h
@@ -61,8 +61,8 @@ typedef PTEntry *PTEntriesPtr;
 #define VA_OFFSET(va) ((u64)(va) & 0xFFF)
 #define PTE_ADDRESS(pte)   ((pte) & ~0xFFFF000000000FFF)
 #define PTE_FLAGS(pte)  ((pte) & 0xFFFF000000000FFF)
-#define P2N(addr) (addr>>12)
-#define PAGE_BASE(addr) (addr & ~(PAGE_SIZE - 1))
+#define P2N(addr) ((addr)>>12)
+#define PAGE_BASE(addr) ((u64)(addr) & ~(PAGE_SIZE - 1))
 
 #define VA_PART0(va) (((u64)(va) & 0xFF8000000000) >> 39)
 #define VA_PART1(va) (((u64)(va) & 0x7FC0000000) >> 30)
